requirements.txt
Django==4.0.1
psycopg2_binary==2.9.3
djangorestframework==3.13.1
django-filter==21.1
djangorestframework-simplejwt==5.0.0
django-cors-headers==4.0.0
drf-nested-routers==0.93.4

pytest-django==4.5.2
pytest-cov==4.1.0